+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_button_press_topdown_wall_v2.py
@@ -25,7 +25,7 @@ class SawyerButtonPressTopdownWallEnvV2(SawyerXYZEnv):
             'obj_init_pos': np.array([0, 0.8, 0.115], dtype=np.float32),
             'hand_init_pos': np.array([0, 0.4, 0.2], dtype=np.float32),
         }
-        self.goal = np.array([0, 0.88, 0.1])
+        self.goal = np.array([0, 0.88, 0.1], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.hand_init_pos = self.init_config['hand_init_pos']
