+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_button_press_v2.py
@@ -24,7 +24,7 @@ class SawyerButtonPressEnvV2(SawyerXYZEnv):
             'obj_init_pos': np.array([0., 0.9, 0.115], dtype=np.float32),
             'hand_init_pos': np.array([0, 0.4, 0.2], dtype=np.float32),
         }
-        self.goal = np.array([0, 0.78, 0.12])
+        self.goal = np.array([0, 0.78, 0.12], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.hand_init_pos = self.init_config['hand_init_pos']
         goal_low = self.hand_low
