+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_coffee_push_v2.py
@@ -26,10 +26,10 @@ class SawyerCoffeePushEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             'obj_init_angle': 0.3,
-            'obj_init_pos': np.array([0., .6, 0.]),
-            'hand_init_pos': np.array([0., .4, .2]),
+            'obj_init_pos': np.array([0., .6, 0.], dtype=np.float32),
+            'hand_init_pos': np.array([0., .4, .2], dtype=np.float32),
         }
-        self.goal = np.array([0., 0.75, 0])
+        self.goal = np.array([0., 0.75, 0], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
