+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_dial_turn_v2.py
@@ -25,10 +25,10 @@ class SawyerDialTurnEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos': np.array([0, 0.7, 0.0]),
+            'obj_init_pos': np.array([0, 0.7, 0.0], dtype=np.float32),
             'hand_init_pos': np.array([0, 0.6, 0.2], dtype=np.float32),
         }
-        self.goal = np.array([0., 0.73, 0.08])
+        self.goal = np.array([0., 0.73, 0.08], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.hand_init_pos = self.init_config['hand_init_pos']
