+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_disassemble_peg_v2.py
@@ -25,10 +25,10 @@ class SawyerNutDisassembleEnvV2(SawyerXYZEnv):
 
         self.init_config = {
             'obj_init_angle': 0.3,
-            'obj_init_pos': np.array([0, 0.7, 0.025]),
+            'obj_init_pos': np.array([0, 0.7, 0.025], dtype=np.float32),
             'hand_init_pos': np.array((0, 0.4, 0.2), dtype=np.float32),
         }
-        self.goal = np.array([0, 0.8, 0.17])
+        self.goal = np.array([0, 0.8, 0.17], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
