+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_door_v2.py
@@ -24,12 +24,12 @@ class SawyerDoorEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_angle': np.array([0.3, ]),
-            'obj_init_pos': np.array([0.1, 0.95, 0.15]),
-            'hand_init_pos': np.array([0, 0.6, 0.2]),
+            'obj_init_angle': np.array([0.3, ], dtype=np.float32),
+            'obj_init_pos': np.array([0.1, 0.95, 0.15], dtype=np.float32),
+            'hand_init_pos': np.array([0, 0.6, 0.2], dtype=np.float32),
         }
 
-        self.goal = np.array([-0.2, 0.7, 0.15])
+        self.goal = np.array([-0.2, 0.7, 0.15], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
