+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_handle_pull_v2.py
@@ -24,10 +24,10 @@ class SawyerHandlePullEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos': np.array([0, 0.9, 0.0]),
-            'hand_init_pos': np.array((0, 0.6, 0.2),),
+            'obj_init_pos': np.array([0, 0.9, 0.0], dtype=np.float32),
+            'hand_init_pos': np.array((0, 0.6, 0.2), dtype=np.float32),
         }
-        self.goal = np.array([0, 0.8, 0.14])
+        self.goal = np.array([0, 0.8, 0.14], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.hand_init_pos = self.init_config['hand_init_pos']
