+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_plate_slide_v2.py
@@ -32,7 +32,7 @@ class SawyerPlateSlideEnvV2(SawyerXYZEnv):
             'obj_init_pos': np.array([0., 0.6, 0.], dtype=np.float32),
             'hand_init_pos': np.array((0, 0.6, 0.2), dtype=np.float32),
         }
-        self.goal = np.array([0., 0.85, 0.02])
+        self.goal = np.array([0., 0.85, 0.02], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
