+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_soccer_v2.py
@@ -29,11 +29,11 @@ class SawyerSoccerEnvV2(SawyerXYZEnv):
         )
 
         self.init_config = {
-            'obj_init_pos': np.array([0, 0.6, 0.03]),
+            'obj_init_pos': np.array([0, 0.6, 0.03], dtype=np.float32),
             'obj_init_angle': 0.3,
-            'hand_init_pos': np.array([0., .6, .2]),
+            'hand_init_pos': np.array([0., .6, .2], dtype=np.float32),
         }
-        self.goal = np.array([0., 0.9, 0.03])
+        self.goal = np.array([0., 0.9, 0.03], dtype=np.float32)
         self.obj_init_pos = self.init_config['obj_init_pos']
         self.obj_init_angle = self.init_config['obj_init_angle']
         self.hand_init_pos = self.init_config['hand_init_pos']
